MIWM-302: Add constants for hosts

+++ src/main/java/miwm/job4me/services/tokens/PasswordResetTokenServiceImpl.java
@@ -1,7 +1,10 @@
 package miwm.job4me.services.tokens;
 
+import miwm.job4me.exceptions.InvalidArgumentException;
+import miwm.job4me.exceptions.NoSuchElementFoundException;
 import miwm.job4me.model.token.PasswordResetToken;
 import miwm.job4me.repositories.users.PasswordTokenRepository;
+import miwm.job4me.validators.fields.IdValidator;
 import org.springframework.stereotype.Service;
 
 import javax.transaction.Transactional;
@@ -12,11 +15,13 @@ import java.util.Set;
 public class PasswordResetTokenServiceImpl implements PasswordResetTokenService {
 
     private final PasswordTokenRepository passwordTokenRepository;
+    private final IdValidator idValidator;
+    private final String ENTITY_TOKEN = "PasswordResetToken";
 
-    public PasswordResetTokenServiceImpl(PasswordTokenRepository passwordTokenRepository) {
+    public PasswordResetTokenServiceImpl(PasswordTokenRepository passwordTokenRepository, IdValidator idValidator) {
         this.passwordTokenRepository = passwordTokenRepository;
+        this.idValidator = idValidator;
     }
-
 
     @Override
     public Set<PasswordResetToken> findAll() {
@@ -25,33 +30,47 @@ public class PasswordResetTokenServiceImpl implements PasswordResetTokenService 
 
     @Override
     public PasswordResetToken findById(Long id) {
+        idValidator.validateLongId(id, ENTITY_TOKEN);
         Optional<PasswordResetToken> token = passwordTokenRepository.findById(id);
         if(token.isPresent())
             return token.get();
         else
-            return null;
+           throw new NoSuchElementFoundException("Reset token with given id does not exist");
     }
 
     @Override
     @Transactional
     public PasswordResetToken save(PasswordResetToken token) {
-        return passwordTokenRepository.save(token);
+        if(token != null)
+            return passwordTokenRepository.save(token);
+        else
+            throw new InvalidArgumentException("Token cannot be null");
     }
 
     @Override
     @Transactional
     public void delete(PasswordResetToken token) {
-        passwordTokenRepository.delete(token);
+        if(token != null)
+            passwordTokenRepository.delete(token);
+        else
+            throw new InvalidArgumentException("Token cannot be null");
     }
 
     @Override
     @Transactional
     public void deleteById(Long id) {
-        passwordTokenRepository.deleteById(id);
+        idValidator.validateLongId(id, ENTITY_TOKEN);
+        if(findById(id) != null)
+            passwordTokenRepository.deleteById(id);
+        else
+            throw new NoSuchElementFoundException("Reset token with given id does not exist");
     }
 
     @Override
     public PasswordResetToken findByToken(String token) {
-        return passwordTokenRepository.findByToken(token);
+        if(token != null)
+            return passwordTokenRepository.findByToken(token);
+        else
+            throw new InvalidArgumentException("Token value cannot be null");
     }
 }
